% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transDT.R
\name{transDT}
\alias{transDT}
\title{A function for quickly transposing data.table objects}
\usage{
transDT(dt, transCol, rowID)
}
\arguments{
\item{dt}{The data.table object you wish to transpose.}

\item{transCol}{The name of the column that you wish to pivot on. Values in this column will become the new colnames.}

\item{rowID}{This will be the name of the new rownames column}
}
\description{
This function allows you transpose data.table objects very rapidly, whilst maintaining control of row and column names.
}
\examples{
transDT()
}
\keyword{data.table}
